// src/pages/storefront-designer/types/designerTypes.ts
export interface AdditionalLinks {
  caption: string;
  link: string;
  save: boolean;
}

/**
 * Shop design configuration
 */
export interface ShopDesignPage {
  templateID?: string;
  backgroundImage?: string;
  backgroundImageSecondary?: string;
  fullWidthHero?: boolean;
  productSectionText?: string;
  backgroundColor?: string;
  backgroundText?: string;
  logo?: string;
  discordURL?: string;
  instagramURL?: string;
  twitterURL?: string;
  facebookURL?: string;
  linkedinURL?: string;
  tiktokURL?: string;
  webURL?: string;
  telegramURL?: string;
  youtubeURL?: string;
  messengerURL?: string;
  headerIcon?: string;
  shopDesign?: {
    fontfamily?: string;
    headerBackground?: string;
    isHeaderFixed?: boolean;
    hiroLayout?: string;
    hiroTextColor?: string;
    bannerLinks?: Array<AdditionalLinks>;
    footerLinks?: Array<AdditionalLinks>;
    productListTitle?: string;
    iconHeaderColor?: string;
    foreground?: string;
    backgroundBody?: string;
    textColorParagraphs?: string;
    isCollectionShown?: boolean;
    isLogoAsFavicon?: boolean;
    faviconURL?: string;
  };
  template_options?: ITemplateOptions;
  launchDate?: string | null;
}